Make tests better by using a real currency that can cause deprecation warnings and defaulting to raise on deprecation
This way tests are more strict and are not silently passing anymore where there would be deprecations in an app that sets a default currency. This is because the gem default NullCurrency allows comparison and arithmetic across currencies. When this looser behaviour needs to be tested we opt-in to it explicitly.