Use string column names for money columns and money currency columns
We use Rails' `read_attribute` and `write_attribute` methods to implement money_column stuff, which, in Shopify, is mostly how we get Money objects to work with. In performance critical money-manipulation code, the money column readers and writers actually get invoked a lot, so, we should make them reasonably fast. It turns out that inside `read_attribute` and `write_attribute`, Rails does a lot of string manipulation to check for aliases and index into the actual attributes storage backend. If we pass in a symbol to those methods, the `#to_s` calls that get invoked super often actually end up allocating a remarkable number of strings. If instead we pass strings right into `read_attribute` and `write_attribute`, all those `#to_s` calls are no-ops, so a lot less string garbage is created. So, this change switches the metaprogramming that defines money column methods (and eventually that calls the Rails methods in question) to define using strings instead of symbols, so strings get passed down. It worked! In the Shopify sales ETL, I saw a 60% reduction of immediately garbage collected strings having to do with money manipulation after this change.