Enable money_column to write currency to DB when default currency is nil (#167)
## What

When `Money`'s default currency is set to `nil`, `money_column`, more specifically `#write_money_attribute`, fails to write the currency provided by the input Money object to the database. Instead it raises `ArgumentError, 'missing currency'`. 

Line 51 is only checking for the currency from one of the following: 
- Money's default currency
- The column's hardcoded currency
- The currency already assigned to the record's corresponding currency column

https://github.com/Shopify/money/blob/df3204ea232a657414f1c5f421d206a50be9dac4/lib/money_column/active_record_hooks.rb#L41-L64

## How

Enable the raw currency source to fall back to the input money object's currency. 
First check the default currency, then the currency already in the column finally, check the input currency. 
Obviously we want to rescue from the input not being a Money object so always default to nil. 

Note: If a `Money.default_currency` is set, options['currency'] will return the default currency so we'll never evaluate the column or input. 

## Why

This enables applications that don't have a default currency set to still write currency to the database with `money_column`.