Update allocate to assign extra pennies more fairly
The current implementation favours the order of the splits too highly. This results in unexpected subunits (pennies)
being assigned to the first ratios, when the decimal value shows a value higher than half a subunit.

For example.

If you split 5 subunits between two parties with 45% to the first party and 55% to the second. You'd expect
the second party to get the extra subunit (55% of 5 is 2.75 vs 45% of 5 is 2.25). But the previous implementation
would only floor the amounts then assign the remaining subunits round robin, this resulted in the first party getting
the extra subunit which is not usually what's you'd expect from a split.