Treat MRU as if it has no subunits.
Stripe does not currently support MRU, but they do support MGA (the other currency in the world that subdivides into fifths).  For MGA, Stripe ignores the subunits, and only deals in full units of the currency.  It seems safest to follow Stripe's lead for MGA when configuring MRU, so this change pretends that there are no subunits of MRU.

Looking at some typical prices in Mauritania, this seems a safe thing to do.  A head of lettuce costs MRU 20, a loaf of bread MRU MRU 235, a dozen eggs MRU 99, and a one-way ticket on local public transit MRU 21... so it's unlikely that anyone will need to price things in fractions smaller than MRU 1.

(At today's mid-market rate, MRU 1 == USD 0.0266212)