✨ Add subunit override for Stripe
Stripe follows the ISO4217 conventions for most currencies, but not for
all, such as ISK. This means that when interacting with the Stripe API
and using subunits/from_subunits the amount is off by 2 orders of
magnitude.

Since ISO4217 is the only real standard defining this conversion ratio,
we keep this as our default behaviour.