Fix money column with non money objects and null default currency
The old behaviour would raise an exception even when a currency was provided
```
MyModel.new(amount: 10, currency: 'USD')
```
This is because when the amount attribute writer is called it does not yet know of the currency. In order to solve this issue, this PR reverts back to supporting legacy non-money objects by saving them as is and not converting them to money objects. In doing so it is possible the DB value to be different than the money value, since no rounding has been applied.