Remove the idea of a 'default' parser
We already had two and introduced two more (with different return semantics). It's better to make an informed choice for every callsite than implicitly relying on global configuration and making a change everywhere.

If a change everywhere is desired for simple apps, it's as simple as replacing 'Money.parse' with 'Money::Parser::Fuzzy.parse' to keep the same behaviour.