Decouple LocaleAwareParser from Action View
Railties allow us to hook into Action View's loading process and do nothing if we're not in a Rails app.
Testing them in isolation is not trivial. The behaviour we setup here however is, and the core logic is already well tested, so skipping over a few lines of glue code is an acceptable tradeoff.