Optimize `Money#split(n)` for large `n`
When passing a very large number to split, it result in allocating
between 2 and 3 arrays of such size. e..g `Money.new(1_000_000).split(1_000_000)`
allocates an array of size 1M, then concat it with an empty array, which cause
Ruby to allocate another array of 1M entries.

All this cause two `8MB` mallocs, which doesn't sound like that much
but if done a few times is likely to trigger Ruby GC.

Instead we can return a small enumerable that will simply act like an
array, but won't actually materialize it.